+++ packages/vseed/src/pipeline/advanced/chart/pipes/encoding/dualAxis.ts
@@ -71,6 +71,10 @@ const generateDimensionEncoding = (dimensions: Dimensions, encoding: Encoding) =
   // tooltip
   encoding.tooltip = unique(dimensions.map((item) => item.id))
   encoding.tooltip = encoding.tooltip.filter((d) => d !== MeasureName)
+
+  // label
+  encoding.label = unique(dimensions.filter((item) => item.encoding === 'label').map((item) => item.id))
+  encoding.label = encoding.label.filter((d) => d !== MeasureName)
 }
 /**
  * --------------------指标--------------------
@@ -96,7 +100,11 @@ const generateMeasureEncoding = (measures: Measures, encoding: Encoding) => {
     encoding.color = color
   }
 
+  // label
+  const label = unique(measures.filter((item) => item.encoding === 'label').map((item) => item.id))
+  encoding.label = unique([...(encoding.label || []), ...label])
+
   // tooltip
   const tooltip = unique(measures.filter((item) => item.encoding === 'tooltip').map((item) => item.id))
-  encoding.tooltip = unique([...(encoding.tooltip || []), ...tooltip])
+  encoding.tooltip = unique([...(encoding.tooltip || []), ...label, ...tooltip])
 }