+++ packages/vseed/src/pipeline/advanced/chart/pipes/measures/deleteTooltipAndLabelMeasure.ts
@@ -1,8 +1,8 @@
 import { deleteMeasureTreeByCallback } from 'src/pipeline/utils'
 import type { AdvancedPipe, Measure } from 'src/types'
 
-export const deleteTooltipMeasure: AdvancedPipe = (advancedVSeed) => {
-  const deleteBy = (measure: Measure) => measure.encoding === 'tooltip'
+export const deleteTooltipAndLabelMeasure: AdvancedPipe = (advancedVSeed) => {
+  const deleteBy = (measure: Measure) => measure.encoding === 'tooltip' || measure.encoding === 'label'
 
   const measureTree = deleteMeasureTreeByCallback(advancedVSeed.measures, deleteBy)